requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
transformers
torch
pillow
librosa
numpy
